+++ src/pages/Nuri/Main/components/Comment.js
@@ -1,24 +1,22 @@
 import { useState } from 'react';
 
-function Comment(props) {
-  const { userName, userComment } = props;
-  const [deleteBox, setDeleteBox] = useState({ display: 'block' });
+function Comment({ userName, userComment, onRemoveClick, id }) {
   const [heart, setHeart] = useState(false);
-  const handleDelete = e => {
-    setDeleteBox({ display: 'none' });
+  const handleRemove = e => {
+    onRemoveClick(id);
   };
-  const onLike = e => {
+  const handleToggleLike = e => {
     setHeart(like => !like);
   };
   return (
-    <div className="Comment" style={deleteBox}>
+    <div className="Comment">
       <li className="add">
         <span className="comment_id">{userName}</span>
         <span className="comment_con">{userComment}</span>
-        <span onClick={onLike} className="comment_heart">
+        <span onClick={handleToggleLike} className="comment_heart">
           <i className={heart ? 'fa-solid fa-heart' : 'fa-regular fa-heart'} />
         </span>
-        <a onClick={handleDelete} href="#!">
+        <a onClick={handleRemove} href="#!">
           삭제
         </a>
       </li>
