+++ src/pages/Nuri/Main/components/Footer.js
@@ -1,0 +1,38 @@
+function Footer() {
+  return (
+    <div className="footer">
+      <ul className="footer_list_wrap">
+        <li>
+          <a href="#!">소개</a>
+        </li>
+        <li>
+          <a href="#!">도움말</a>
+        </li>
+        <li>
+          <a href="#!">홍보 센터</a>
+        </li>
+        <li>
+          <a href="#!">API</a>
+        </li>
+        <li>
+          <a href="#!">채용 정보</a>
+        </li>
+        <li>
+          <a href="#!">개인정보처리방침</a>
+        </li>
+        <li>
+          <a href="#!">약관</a>
+        </li>
+        <li>
+          <a href="#!">위치</a>
+        </li>
+        <li>
+          <a href="#!">언어</a>
+        </li>
+      </ul>
+      <p>© 2022 INSTAGRAM FROM META</p>
+    </div>
+  );
+}
+
+export default Footer;
